src/main/java/kr/megaptera/jdbc/assignment/application/GetPostsService.java
package kr.megaptera.jdbc.assignment.application;

import kr.megaptera.jdbc.assignment.daos.PostDao;
import kr.megaptera.jdbc.assignment.domain.Post;
import kr.megaptera.jdbc.assignment.dtos.PostDto;
import kr.megaptera.jdbc.assignment.repositories.PostRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;

@Service
@RequiredArgsConstructor
public class GetPostsService {

    private final PostDao postDao;

    public List<PostDto> getPostList() {
        List<Post> posts = postDao.findAll();

        return posts.stream().map(PostDto::new).toList();
    }
}
